Booksinfo_bot
This app is for searching books by query (title, author, etc.). The list of books is displayed with all information about them.
Features: user can enter a search query in the input field;
the user can send a request that accesses an API that returns an array with books and additional information (title, author, publication date, etc.).